ml-services/recommendation-engine/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
scikit-learn==1.8.0
numpy==2.3.2
joblib==1.5.1
requests==2.32.4
apscheduler==3.10.4
croniter==6.0.0
python-dotenv>=1.0.1
